Refactor db entities and created entities migration
