+++ package.json
@@ -4,6 +4,7 @@
     "description": "Flixxo's technical task",
     "main": "dist/index.js",
     "scripts": {
+        "typeorm": "ts-node-dev ./node_modules/typeorm/cli.js",
         "prebuild": "tslint -c tslint.json -p tsconfig.json --fix",
         "compile": "tsc && node dist/index.js",
         "dev": "./node_modules/nodemon/bin/nodemon.js -e ts  --exec \"npm run compile\"",
