+++ src/db/DataSource.ts
@@ -1,4 +1,4 @@
-import { AppDataSource } from "./config/data-source";
+import AppDataSource from "../data-source";
 
 
 export class DataSource {
