+++ tsconfig.json
@@ -17,5 +17,5 @@
   },
   "include": [
       "src/**/*"
-  ]
+  ],
 }
